Added all Controllers

+++ src/main/java/com/pavelchak/service/CityService.java
@@ -1,7 +1,9 @@
 package com.pavelchak.service;
 
 import com.pavelchak.Repository.CityRepository;
+import com.pavelchak.Repository.PersonRepository;
 import com.pavelchak.domain.CityEntity;
+import com.pavelchak.exceptions.ExistsPersonsForCityException;
 import com.pavelchak.exceptions.NoSuchCityException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -14,6 +16,9 @@ public class CityService {
     @Autowired
     CityRepository cityRepository;
     private boolean ascending;
+
+    @Autowired
+    PersonRepository personRepository;
 
     public List<CityEntity> getAllCity(){
         return cityRepository.findAll();
@@ -39,9 +44,10 @@ public class CityService {
     }
 
     @Transactional
-    public void deleteCity(Long city_id) throws NoSuchCityException {
+    public void deleteCity(Long city_id) throws NoSuchCityException, ExistsPersonsForCityException {
         CityEntity cityEntity=cityRepository.findOne(city_id);
         if(cityEntity==null) throw new NoSuchCityException();
+        if(cityEntity.getPersons().size()!=0) throw new ExistsPersonsForCityException();
         cityRepository.delete(cityEntity);
     }
